+++ lib/strong_migrations.rb
@@ -29,7 +29,7 @@ module StrongMigrations
       :target_postgresql_version, :target_mysql_version, :target_mariadb_version,
       :enabled_checks, :lock_timeout, :statement_timeout, :check_down, :target_version,
       :safe_by_default, :target_sql_mode, :lock_timeout_retries, :lock_timeout_retry_delay,
-      :alphabetize_schema, :skip_databases, :remove_invalid_indexes
+      :alphabetize_schema, :skipped_databases, :remove_invalid_indexes
     attr_writer :lock_timeout_limit
   end
   self.auto_analyze = false
@@ -40,7 +40,7 @@ module StrongMigrations
   self.safe_by_default = false
   self.check_down = false
   self.alphabetize_schema = false
-  self.skip_databases = []
+  self.skipped_databases = []
   self.remove_invalid_indexes = false
 
   # private
@@ -84,6 +84,10 @@ module StrongMigrations
     else
       false
     end
+  end
+
+  def self.skip_database(database)
+    self.skipped_databases << database
   end
 end