+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 2.1.0 (unreleased)
 
-- Added `skip_databases` option
+- Added `skip_database` method
 - Added experimental `remove_invalid_indexes` option
 - Added warning for unsupported adapters
 - Improved output for `db:forward`, `db:rollback`, `db:migrate:up`, and `db:migrate:down`